requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.1
DAWG-Python==0.7.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
executing==1.2.0
idna==3.4
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymystem3==0.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
